refactor(transport): extract DDS publisher path into its own file
Moves createPublisher, appendPublisher, takeAllPublishers, and the
DDSTransportPublisherImpl concrete type into a focused extension file.
DDSTransportPublisherImpl stays default-internal because the core file's
'publishers' dictionary references it by name.

The DDSTransportSession.swift residue is now ~150 lines: properties,
init, open, close, checkHealth, bridgeQoS, generateFallbackSessionId.
